chore: deep analysis fixes, extended E2E tests, and git hygiene
- Fixed uninitialized ScrollPosition and Cursor members in EditorDocument
- Fixed OperationalTransform insert-vs-delete logic bug
- Fixed compiler warnings (signed/unsigned, unused vars, volatile assignment)
- Added 52 new E2E tests covering 11 previously untested components
- Removed 23 tracked binaries and archives from git
- Updated .gitignore with comprehensive patterns
- Updated SKILL.md with latest test counts (97/97 passing)

+++ include/bolt/editor/syntax_highlighter.hpp
@@ -65,7 +65,7 @@ public:
             for (const auto& rule : rules) {
                 std::smatch match;
                 if (std::regex_search(remaining, match, rule.first, std::regex_constants::match_continuous)) {
-                    if (match.position(0) < bestMatch) {
+                    if (static_cast<size_t>(match.position(0)) < bestMatch) {
                         bestMatch = match.position(0);
                         bestType = rule.second;
                         matchedText = match.str(0);
